+++ .vscode/settings.json
@@ -1,5 +1,4 @@
 {
-    "python.analysis.typeCheckingMode": "standard",
     "cSpell.words": [
         "argmin",
         "argsorted",
@@ -47,6 +46,11 @@
         "thiserror",
         "unpadded"
     ],
+    "terminal.integrated.env.linux": {
+        "PYTHONPATH": "${workspaceFolder}/src"
+    },
+    "python.analysis.typeCheckingMode": "standard",
+    "python.analysis.extraPaths": ["src"],
     "python.testing.pytestArgs": [
         "tests"
     ],
