+++ pyproject.toml
@@ -25,5 +25,5 @@ dev = [
 
 [tool.pytest.ini_options]
 pythonpath = [
-    ".",
+    ".", "src",
 ]
