+++ scripts/train_model.py
@@ -1,9 +1,9 @@
 from argparse import ArgumentParser
 from pathlib import Path
 
-from src.config_loader import load_yaml_file
-from src.training.train import train_model
-from src.training.training_config import TrainingConfig
+from rtmdet_object_detection_dev.config_loader import load_yaml_file
+from rtmdet_object_detection_dev.training.train import train_model
+from rtmdet_object_detection_dev.training.training_config import TrainingConfig
 
 
 def main() -> None:
