+++ src/rtmdet_object_detection_dev/losses/loss_fn_factory.py
@@ -1,6 +1,6 @@
 import torch.nn as nn
 
-from src.losses.rtmdet_loss import RTMDetLoss
+from rtmdet_object_detection_dev.losses.rtmdet_loss import RTMDetLoss
 
 
 def get_loss_fn(name: str, config: dict) -> nn.Module:
