+++ src/rtmdet_object_detection_dev/model/rtmdet_models.py
@@ -4,7 +4,7 @@ import torch.nn as nn
 from .backbone.cspnext import CSPNeXt
 from .neck.cspnext_pafpn import CSPNeXtPAFPN
 from .head.rtmdet_sep_bn_head import RTMDetSepBNHead
-from src.dataclasses.rtmdet_output import RTMDetOutput
+from rtmdet_object_detection_dev.dataclasses.rtmdet_output import RTMDetOutput
 
 
 class RTMDet(nn.Module):
